requirements.txt
fire==0.5.0
google.generativeai==0.8.3
numpy==1.26.4
openai==1.59.7
pandas==2.2.3
pydantic==2.10.6
sympy==1.13.1
transformers==4.48.1
vllm==0.6.6
